Moved all database read functionality to new class
Keeping database reading functionality scattered across the application
would have made the application more difficult to maintain later.

In making this change, I kept all table names stored in readFromDatabase.php -
it would not be clear to an unfamiliar programmer that each
LogsheetComponent class name needed to match the name of the table where
its entries originated from, and each field of each LogsheetComponent
object needed to match the name of the column its entries originated from.

Also, each LogsheetComponent attribute is now stored in its own field of the
object instead of within a single attribute array field. It was not
always clear the type of object that was stored in each "slot" of
the array (e.g. an id to be used to construct a LogsheetComponent object
vs. the LogsheetComponent itself). I also made this change to enable
readFromDatabase.php to completely build LogsheetComponent objects on
its own (in case the way an object is represented in the database changes
in future, only one method in readFromDatabase would need to be changed).