Save prerecord date as DATE, not DATETIME
The prerecord date as entered by the user is a date
without any time assoicated with it. The system needed to
add an arbitrary time to this prerecord date to match the DATETIME
type in the database. Adding this arbitrary time is misleading -
the user never intended it to be there, and the newly formed
datetime itself could be misread in future.